+++ migrations/0007_create_images.up.sql
@@ -4,15 +4,15 @@ CREATE TABLE images(
   updated_at timestamp without time zone not null,
   name varchar not null,
   type varchar not null,
-  size integer not null,
+  size int not null,
   url varchar not null,
-  property_id integer references properties,
-  unit_id integer references units check (property_id is not null or unit_id is not null),
+  property_id bigint references properties,
+  unit_id bigint references units check (property_id is not null or unit_id is not null),
   uploaded boolean default false,
   position smallint
 );
 
-ALTER TABLE units ADD floor_plan_image_id int references images;
+ALTER TABLE units ADD floor_plan_image_id bigint references images;
 
 CREATE INDEX images_property_id ON images(property_id);
 CREATE INDEX images_unit_id ON images(unit_id);
